Add more advanced "random' number use for loadfiles. Make it possible to use random number and add alignment, modulo and offset to them so it is possible in nfs for example to specify offset as */0x1000%5000000 meaning : random number between 0 and 500000 aligned to 4096
update the nfs backend to accept 64 bit offsets

+++ dbench.h
@@ -170,7 +170,7 @@ struct dbench_op {
 	const char *fname;
 	const char *fname2;
 	const char *status;
-	int params[10];
+	int64_t params[10];
 };
 
 struct backend_op {
